+++ CHANGELOG.md
@@ -74,6 +74,9 @@
 * Glitchy rendering when scrolling in the scrollback, on compositors
   that does not allow Wayland buffer re-use (e.g. KDE/plasma)
   ([#1173][1173])
+* Scrollback search matches not being highlighted correctly, on
+  compositors that does now allow Wayland buffer re-use
+  (e.g. KDE/plasma).
 
 [1173]: https://codeberg.org/dnkl/foot/issues/1173